Network.Wai.Handler.Warp

Description

A fast, light-weight HTTP server handler for WAI.

Run a Warp server

run :: Port -> Application -> IO ()

Run an Application on the given port. This calls runSettings with defaultSettings.

runEnv :: Port -> Application -> IO ()

Run an Application on the port present in the PORT environment variable

Uses the Port given when the variable is unset.

Since 3.0.9

runSettings :: Settings -> Application -> IO ()

Run an Application with the given Settings.

runSettingsSocket :: Settings -> Socket -> Application -> IO ()

Same as runSettings, but uses a user-supplied socket instead of opening one. This allows the user to provide, for example, Unix named socket, which can be used when reverse HTTP proxying into your application.

Note that the settingsPort will still be passed to Applications via the serverPort record.

When the listen socket in the second argument is closed, all live connections are gracefully shut down.

runSettingsConnection :: Settings -> IO (Connection, SockAddr) -> Application -> IO ()

Allows you to provide a function which will return a Connection. In cases where creating the Connection can be expensive, this allows the expensive computations to be performed in a separate thread instead of the main server loop.

Since 1.3.5

runSettingsConnectionMaker :: Settings -> IO (IO Connection, SockAddr) -> Application -> IO ()

runSettingsConnectionMakerSecure :: Settings -> IO (IO (Connection, Transport), SockAddr) -> Application -> IO ()

Allows you to provide a function which will return a function which will return Connection.

Since 2.1.4

Settings

data Settings

Various Warp server settings. This is purposely kept as an abstract data type so that new settings can be added without breaking backwards compatibility. In order to create a Settings value, use defaultSettings and the various 'set' functions to modify individual fields. For example:

setTimeout 20 defaultSettings

defaultSettings :: Settings

The default settings for the Warp server. See the individual settings for the default value.

Setters

setPort :: Int -> Settings -> Settings

Port to listen on. Default value: 3000

Since 2.1.0

setHost :: HostPreference -> Settings -> Settings

Interface to bind to. Default value: HostIPv4

Since 2.1.0

setOnException :: (Maybe Request -> SomeException -> IO ()) -> Settings -> Settings

What to do with exceptions thrown by either the application or server. Default: ignore server-generated exceptions (see InvalidRequest) and print application-generated exceptions to stderr.

Since 2.1.0

setOnExceptionResponse :: (SomeException -> Response) -> Settings -> Settings

A function to create a Response when an exception occurs.

Default: 500, text/plain, "Something went wrong"

Since 2.1.0

setOnOpen :: (SockAddr -> IO Bool) -> Settings -> Settings

What to do when a connection is opened. When False is returned, the connection is closed immediately. Otherwise, the connection is going on. Default: always returns True.

Since 2.1.0

setOnClose :: (SockAddr -> IO ()) -> Settings -> Settings

What to do when a connection is closed. Default: do nothing.

Since 2.1.0

setTimeout :: Int -> Settings -> Settings

Timeout value in seconds. Default value: 30

Since 2.1.0

setManager :: Manager -> Settings -> Settings

Use an existing timeout manager instead of spawning a new one. If used, settingsTimeout is ignored.

Since 2.1.0

setFdCacheDuration :: Int -> Settings -> Settings

Cache duration time of file descriptors in seconds. 0 means that the cache mechanism is not used. Default value: 10

setBeforeMainLoop :: IO () -> Settings -> Settings

Code to run after the listening socket is ready but before entering the main event loop. Useful for signaling to tests that they can start running, or to drop permissions after binding to a restricted port.

Default: do nothing.

Since 2.1.0

setNoParsePath :: Bool -> Settings -> Settings

Perform no parsing on the rawPathInfo.

This is useful for writing HTTP proxies.

Default: False

Since 2.1.0

setInstallShutdownHandler :: (IO () -> IO ()) -> Settings -> Settings

A code to install shutdown handler.

For instance, this code should set up a UNIX signal handler. The handler should call the first argument, which close the listen socket, at shutdown.

Default: does not install any code.

Since 3.0.1

setServerName :: ByteString -> Settings -> Settings

Default server name if application does not set one.

Since 3.0.2

setMaximumBodyFlush :: Maybe Int -> Settings -> Settings

The maximum number of bytes to flush from an unconsumed request body.

By default, Warp does not flush the request body so that, if a large body is present, the connection is simply terminated instead of wasting time and bandwidth on transmitting it. However, some clients do not deal with that situation well. You can either change this setting to Nothing to flush the entire body in all cases, or in your application ensure that you always consume the entire request body.

Default: 8192 bytes.

Since 3.0.3

setFork :: (((forall a. IO a -> IO a) -> IO ()) -> IO ()) -> Settings -> Settings

Code to fork a new thread to accept a connection.

This may be useful if you need OS bound threads, or if you wish to develop an alternative threading model.

Default: void . forkIOWithUnmask

Since 3.0.4

setProxyProtocolNone :: Settings -> Settings

Do not use the PROXY protocol.

Since 3.0.5

setProxyProtocolRequired :: Settings -> Settings

Require PROXY header.

This is for cases where a "dumb" TCP/SSL proxy is being used, which cannot add an X-Forwarded-For HTTP header field but has enabled support for the PROXY protocol.

See http://www.haproxy.org/download/1.5/doc/proxy-protocol.txt and http://docs.aws.amazon.com/ElasticLoadBalancing/latest/DeveloperGuide/TerminologyandKeyConcepts.html#proxy-protocol.

Only the human-readable header format (version 1) is supported. The binary header format (version 2) is not supported.

Since 3.0.5

setProxyProtocolOptional :: Settings -> Settings

Use the PROXY header if it exists, but also accept connections without the header. See setProxyProtocolRequired.

WARNING: This is contrary to the PROXY protocol specification and using it can indicate a security problem with your architecture if the web server is directly accessable to the public, since it would allow easy IP address spoofing. However, it can be useful in some cases, such as if a load balancer health check uses regular HTTP without the PROXY header, but proxied connections do include the PROXY header.

Since 3.0.5

Getters

getPort :: Settings -> Int

Get the listening port.

Since 2.1.1

getHost :: Settings -> HostPreference

Get the interface to bind to.

Since 2.1.1

Accessors

Note: these accessors are deprecated, please use the set versions instead.

settingsPort :: Settings -> Int

Deprecated: Use setPort instead

Port to listen on. Default value: 3000

settingsHost :: Settings -> HostPreference

Deprecated: Use setHost instead

Default value: HostIPv4

settingsOnException :: Settings -> Maybe Request -> SomeException -> IO ()

Deprecated: Use setOnException instead

What to do with exceptions thrown by either the application or server. Default: ignore server-generated exceptions (see InvalidRequest) and print application-generated applications to stderr.

settingsOnExceptionResponse :: Settings -> SomeException -> Response

Deprecated: Use setOnExceptionResponse instead

A function to create Response when an exception occurs.

Default: 500, text/plain, "Something went wrong"

Since 2.0.3

settingsOnOpen :: Settings -> SockAddr -> IO Bool

Deprecated: Use setOnOpen instead

What to do when a connection is open. When False is returned, the connection is closed immediately. Otherwise, the connection is going on. Default: always returns True.

settingsOnClose :: Settings -> SockAddr -> IO ()

Deprecated: Use setOnClose instead

What to do when a connection is close. Default: do nothing.

settingsTimeout :: Settings -> Int

Deprecated: Use setTimeout instead

Timeout value in seconds. Default value: 30

settingsManager :: Settings -> Maybe Manager

Deprecated: Use setManager instead

Use an existing timeout manager instead of spawning a new one. If used, settingsTimeout is ignored. Default is Nothing

settingsFdCacheDuration :: Settings -> Int

Deprecated: Use setFdCacheDuration instead

Cache duratoin time of file descriptors in seconds. 0 means that the cache mechanism is not used. Default value: 10

settingsBeforeMainLoop :: Settings -> IO ()

Deprecated: Use setBeforeMainLoop instead

Code to run after the listening socket is ready but before entering the main event loop. Useful for signaling to tests that they can start running, or to drop permissions after binding to a restricted port.

Default: do nothing.

Since 1.3.6

settingsNoParsePath :: Settings -> Bool

Deprecated: Use setNoParsePath instead

Perform no parsing on the rawPathInfo.

This is useful for writing HTTP proxies.

Default: False

Since 2.0.3

Debugging

exceptionResponseForDebug :: SomeException -> Response

Default implementation of settingsOnExceptionResponse for the debugging purpose. 500, text/plain, a showed exception.

defaultShouldDisplayException :: SomeException -> Bool

Apply the logic provided by defaultExceptionHandler to determine if an exception should be shown or not. The goal is to hide exceptions which occur under the normal course of the web server running.

Since 2.1.3

Data types

data Transport

What kind of transport is used for this connection?

Constructors

TCP
Plain channel: TCP

TLS
Encrypted channel: TLS or SSL

Fields

tlsMajorVersion :: Int

tlsMinorVersion :: Int

tlsNegotiatedProtocol :: Maybe ByteString

The result of Application Layer Protocol Negociation in RFC 7301

tlsChiperID :: Word16

data HostPreference :: *

Which host to bind.

Note: The IsString instance recognizes the following special values:

* means HostAny

*4 means HostIPv4

!4 means HostIPv4Only

*6 means HostIPv6

!6 means HostIPv6Only

Any other values is treated as a hostname. As an example, to bind to the IPv4 local host only, use "127.0.0.1".

type Port = Int

TCP port number.

data InvalidRequest

Error types for bad Request.

Constructors

NotEnoughLines [String]

BadFirstLine String

NonHttp

IncompleteHeaders

ConnectionClosedByPeer

OverLargeHeader

BadProxyHeader String

data ConnSendFileOverride

Whether or not ConnSendFileOverride in Connection can be overridden. This is a kind of hack to keep the signature of Connection clean.

Constructors

NotOverride
Don't override

Override Socket
Override with this Socket

Connection

data Connection

Data type to manipulate IO actions for connections.

Constructors

Connection

Fields

connSendMany :: [ByteString] -> IO ()

connSendAll :: ByteString -> IO ()

connSendFile :: FilePath -> Integer -> Integer -> IO () -> [ByteString] -> IO ()

filepath, offset, length, hook action, HTTP headers

connClose :: IO ()

connRecv :: IO ByteString

connReadBuffer :: Buffer

connWriteBuffer :: Buffer

connBufferSize :: BufSize

connSendFileOverride :: ConnSendFileOverride

socketConnection :: Socket -> IO Connection

Default action value for Connection.

Internal

Version

warpVersion :: String

The version of Warp.

Data types

data InternalInfo

Internal information.

Constructors

InternalInfo

Fields

threadHandle :: Handle

fdCacher :: Maybe MutableFdCache

dateCacher :: DateCache

type HeaderValue = ByteString

The type for header value used with HeaderName.

type IndexedHeader = Array Int (Maybe HeaderValue)

Array for a set of HTTP headers.

requestMaxIndex :: Int

The size for IndexedHeader for HTTP Request. From 0 to this corresponds to "Content-Length", "Transfer-Encoding", "Expect", "Connection", "Range", and "Host".

Time out manager

module Network.Wai.Handler.Warp.Timeout

File descriptor cache

withFdCache :: Int -> (Maybe MutableFdCache -> IO a) -> IO a

Creating MutableFdCache and executing the action in the second argument. The first argument is a cache duration in second.

getFd :: MutableFdCache -> FilePath -> IO (Fd, Refresh)

Getting Fd and Refresh from the mutable Fd cacher.

type MutableFdCache = Reaper FdCache (Hash, FdEntry)

Mutable Fd cacher.

type Refresh = IO ()

An action to activate a Fd cache entry.

Date

withDateCache :: (DateCache -> IO a) -> IO a

Creating DateCache and executing the action.

getDate :: DateCache -> IO GMTDate

Getting GMTDate based on DateCache.

type DateCache = IO GMTDate

The type of the cache of the Date header value.

type GMTDate = ByteString

The type of the Date header value.

Request and response

recvRequest

Arguments

:: Settings

-> Connection

-> InternalInfo

-> SockAddr
Peer's address.

-> Source
Where HTTP request comes from.

-> IO (Request, Maybe (IORef Int), IndexedHeader)
Request passed to Application, how many bytes remain to be consumed, if known IndexedHeader of HTTP request for internal use,

Receiving a HTTP request from Connection and parsing its header to create Request.

sendResponse

Arguments

:: ByteString
default server value

-> Connection

-> InternalInfo

-> Request
HTTP request.

-> IndexedHeader
Indexed header of HTTP request.

-> IO ByteString
source from client, for raw response

-> Response
HTTP response including status code and response header.

-> IO Bool
Returing True if the connection is persistent.

Sending a HTTP response to Connection according to Response.

Applications/middlewares MUST specify a proper ResponseHeaders. so that inconsistency does not happen. No header is deleted by this function.

Especially, Applications/middlewares MUST take care of Content-Length, Content-Range, and Transfer-Encoding because they are inserted, when necessary, regardless they already exist. This function does not insert Content-Encoding. It's middleware's responsibility.

The Date and Server header is added if not exist in HTTP response header.

There are three basic APIs to create Response:

responseFile :: Status -> ResponseHeaders -> FilePath -> Maybe FilePart -> Response

HTTP response body is sent by sendfile(). Applications are categorized into simple and sophisticated. Simple applications should specify Nothing to Maybe FilePart. The size of the specified file is obtained by disk access. Then Range is handled. Sophisticated applications should specify Just to Maybe FilePart. They should treat Range (and If-Range) by thierselves. In both cases, Content-Length and Content-Range (if necessary) are automatically added into the HTTP response header. If Content-Length and Content-Range exist in the HTTP response header, they would cause inconsistency. Status is also changed to 206 if necessary.

responseBuilder :: Status -> ResponseHeaders -> Builder -> Response

HTTP response body is created from Builder. Transfer-Encoding: chunked is used in HTTP/1.1.

responseStream :: Status -> ResponseHeaders -> StreamingBody -> Response

HTTP response body is created from Builder. Transfer-Encoding: chunked is used in HTTP/1.1.

responseRaw :: (IO ByteString -> (ByteString -> IO ()) -> IO ()) -> Response -> Response

No header is added and no Transfer-Encoding: is applied.