Network.Wai.Handler.Warp.Internal

Settings

data Settings

Various Warp server settings. This is purposely kept as an abstract data type so that new settings can be added without breaking backwards compatibility. In order to create a Settings value, use defaultSettings and the various 'set' functions to modify individual fields. For example:

setTimeout 20 defaultSettings

Constructors

Settings

Fields

settingsPort :: Port

Port to listen on. Default value: 3000

settingsHost :: HostPreference

Default value: HostIPv4

settingsOnException :: Maybe Request -> SomeException -> IO ()

What to do with exceptions thrown by either the application or server. Default: ignore server-generated exceptions (see InvalidRequest) and print application-generated applications to stderr.

settingsOnExceptionResponse :: SomeException -> Response

A function to create Response when an exception occurs.

Default: 500, text/plain, "Something went wrong"

Since 2.0.3

settingsOnOpen :: SockAddr -> IO Bool

What to do when a connection is open. When False is returned, the connection is closed immediately. Otherwise, the connection is going on. Default: always returns True.

settingsOnClose :: SockAddr -> IO ()

What to do when a connection is close. Default: do nothing.

settingsTimeout :: Int

Timeout value in seconds. Default value: 30

settingsManager :: Maybe Manager

Use an existing timeout manager instead of spawning a new one. If used, settingsTimeout is ignored. Default is Nothing

settingsFdCacheDuration :: Int

Cache duration time of file descriptors in seconds. 0 means that the cache mechanism is not used. Default value: 0

settingsBeforeMainLoop :: IO ()

Code to run after the listening socket is ready but before entering the main event loop. Useful for signaling to tests that they can start running, or to drop permissions after binding to a restricted port.

Default: do nothing.

Since 1.3.6

settingsFork :: ((forall a. IO a -> IO a) -> IO ()) -> IO ()

Code to fork a new thread to accept a connection.

This may be useful if you need OS bound threads, or if you wish to develop an alternative threading model.

Default: void . forkIOWithUnmask

Since 3.0.4

settingsNoParsePath :: Bool

Perform no parsing on the rawPathInfo.

This is useful for writing HTTP proxies.

Default: False

Since 2.0.3

settingsInstallShutdownHandler :: IO () -> IO ()

settingsServerName :: ByteString

Default server name if application does not set one.

Since 3.0.2

settingsMaximumBodyFlush :: Maybe Int

See setMaximumBodyFlush.

Since 3.0.3

settingsProxyProtocol :: ProxyProtocol

Specify usage of the PROXY protocol.

Since 3.0.5.

settingsSlowlorisSize :: Int

Size of bytes read to prevent Slowloris protection. Default value: 2048

Since 3.1.2.

data ProxyProtocol

Specify usage of the PROXY protocol.

Constructors

ProxyProtocolNone
See setProxyProtocolNone.

ProxyProtocolRequired
See setProxyProtocolRequired.

ProxyProtocolOptional
See setProxyProtocolOptional.

Low level run functions

runSettingsConnection :: Settings -> IO (Connection, SockAddr) -> Application -> IO ()

The connection setup action would be expensive. A good example is initialization of TLS. So, this converts the connection setup action to the connection maker which will be executed after forking a new worker thread. Then this calls runSettingsConnectionMaker with the connection maker. This allows the expensive computations to be performed in a separate worker thread instead of the main server loop.

Since 1.3.5

runSettingsConnectionMaker :: Settings -> IO (IO Connection, SockAddr) -> Application -> IO ()

This modifies the connection maker so that it returns TCP for Transport (i.e. plain HTTP) then calls runSettingsConnectionMakerSecure.

runSettingsConnectionMakerSecure :: Settings -> IO (IO (Connection, Transport), SockAddr) -> Application -> IO ()

The core run function which takes Settings, a connection maker and Application. The connection maker can return a connection of either plain HTTP or HTTP over TLS.

Since 2.1.4

data Transport

What kind of transport is used for this connection?

Constructors

TCP
Plain channel: TCP

TLS
Encrypted channel: TLS or SSL

Fields

tlsMajorVersion :: Int

tlsMinorVersion :: Int

tlsNegotiatedProtocol :: Maybe ByteString

The result of Application Layer Protocol Negociation in RFC 7301

tlsChiperID :: Word16

Connection

data Connection

Data type to manipulate IO actions for connections. This is used to abstract IO actions for plain HTTP and HTTP over TLS.

Constructors

Connection

Fields

connSendMany :: [ByteString] -> IO ()

This is not used at this moment.

connSendAll :: ByteString -> IO ()

The sending function.

connSendFile :: SendFile

The sending function for files in HTTP/1.1.

connClose :: IO ()

The connection closing function.

connRecv :: Recv

The connection receiving function. This returns "" for EOF.

connRecvBuf :: RecvBuf

The connection receiving function. This tries to fill the buffer. This returns when the buffer is filled or reaches EOF.

connWriteBuffer :: Buffer

The write buffer.

connBufferSize :: BufSize

The size of the write buffer.

socketConnection :: Socket -> IO Connection

Creating Connection for plain HTTP based on a given socket.

Receive

type Recv = IO ByteString

Type for the action to receive input data

type RecvBuf = Buffer -> BufSize -> IO Bool

Type for the action to receive input data with a buffer. The result boolean indicates whether or not the buffer is fully filled.

makePlainReceiveN :: Socket -> ByteString -> IO (BufSize -> IO ByteString)

This function returns a receiving function based on two receiving functions. The returned function efficiently manages received data which is initialized by the first argument. The returned function may allocate a byte string with malloc().

Buffer

type Buffer = Ptr Word8

Type for buffer

type BufSize = Int

Type for buffer size

bufferSize :: BufSize

The default size of the write buffer: 16384 (2^14 = 1024 * 16). This is the maximum size of TLS record. This is also the maximum size of HTTP/2 frame payload (excluding frame header).

allocateBuffer :: Int -> IO Buffer

Allocating a buffer with malloc().

freeBuffer :: Buffer -> IO ()

Releasing a buffer with free().

copy :: Buffer -> ByteString -> IO Buffer

Copying the bytestring to the buffer. This function returns the point where the next copy should start.

Sendfile

data FileId

Data type to abstract file identifiers. On Unix, a file descriptor would be specified to make use of the file descriptor cache.

Since: 3.1.0

Constructors

FileId

Fields

fileIdPath :: FilePath

fileIdFd :: Maybe Fd

type SendFile = FileId -> Integer -> Integer -> IO () -> [ByteString] -> IO ()

fileid, offset, length, hook action, HTTP headers

Since: 3.1.0

sendFile :: Socket -> Buffer -> BufSize -> (ByteString -> IO ()) -> SendFile

Function to send a file based on sendfile() for Linux/Mac/FreeBSD. This makes use of the file descriptor cache. For other OSes, this is identical to readSendFile.

Since: 3.1.0

readSendFile :: Buffer -> BufSize -> (ByteString -> IO ()) -> SendFile

Function to send a file based on pread()/send() for Unix. This makes use of the file descriptor cache. For Windows, this is emulated by Handle.

Since: 3.1.0

Version

warpVersion :: String

The version of Warp.

Data types

data InternalInfo

Internal information.

Constructors

InternalInfo

Fields

threadHandle :: Handle

timeoutManager :: Manager

fdCacher :: Maybe MutableFdCache

dateCacher :: DateCache

type HeaderValue = ByteString

The type for header value used with HeaderName.

type IndexedHeader = Array Int (Maybe HeaderValue)

Array for a set of HTTP headers.

requestMaxIndex :: Int

The size for IndexedHeader for HTTP Request. From 0 to this corresponds to "Content-Length", "Transfer-Encoding", "Expect", "Connection", "Range", and "Host".

Time out manager

In order to provide slowloris protection, Warp provides timeout handlers. We follow these rules:

A timeout is created when a connection is opened.

When all request headers are read, the timeout is tickled.

Every time at least the slowloris size settings number of bytes of the request body are read, the timeout is tickled.

The timeout is paused while executing user code. This will apply to both the application itself, and a ResponseSource response. The timeout is resumed as soon as we return from user code.

Every time data is successfully sent to the client, the timeout is tickled.

Types

type Manager = Reaper [Handle] Handle

A timeout manager

type TimeoutAction = IO ()

An action to be performed on timeout.

data Handle

A handle used by Manager

Manager

initialize :: Int -> IO Manager

Creating timeout manager which works every N micro seconds where N is the first argument.

stopManager :: Manager -> IO ()

Stopping timeout manager.

withManager

Arguments

:: Int
timeout in microseconds

-> (Manager -> IO a)

-> IO a

Call the inner function with a timeout manager.

Registration

register :: Manager -> TimeoutAction -> IO Handle

Registering a timeout action.

registerKillThread :: Manager -> IO Handle

Registering a timeout action of killing this thread.

Control

tickle :: Handle -> IO ()

Setting the state to active. Manager turns active to inactive repeatedly.

cancel :: Handle -> IO ()

Setting the state to canceled. Manager eventually removes this without timeout action.

pause :: Handle -> IO ()

Setting the state to paused. Manager does not change the value.

resume :: Handle -> IO ()

Setting the paused state to active. This is an alias to tickle.

Exceptions

data TimeoutThread

Constructors

TimeoutThread

File descriptor cache

withFdCache :: Int -> (Maybe MutableFdCache -> IO a) -> IO a

Creating MutableFdCache and executing the action in the second argument. The first argument is a cache duration in second.

getFd :: MutableFdCache -> FilePath -> IO (Fd, Refresh)

Getting Fd and Refresh from the mutable Fd cacher.

type MutableFdCache = Reaper FdCache (Hash, FdEntry)

Mutable Fd cacher.

type Refresh = IO ()

An action to activate a Fd cache entry.

Date

withDateCache :: (DateCache -> IO a) -> IO a

Creating DateCache and executing the action.

getDate :: DateCache -> IO GMTDate

Getting GMTDate based on DateCache.

type DateCache = IO GMTDate

The type of the cache of the Date header value.

type GMTDate = ByteString

The type of the Date header value.

Request and response

data Source

Type for input streaming.

recvRequest

Arguments

:: Settings

-> Connection

-> InternalInfo

-> SockAddr
Peer's address.

-> Source
Where HTTP request comes from.

-> IO (Request, Maybe (IORef Int), IndexedHeader, IO ByteString)
Request passed to Application, how many bytes remain to be consumed, if known IndexedHeader of HTTP request for internal use, Body producing action used for flushing the request body

Receiving a HTTP request from Connection and parsing its header to create Request.

sendResponse

Arguments

:: ByteString
default server value

-> Connection

-> InternalInfo

-> Request
HTTP request.

-> IndexedHeader
Indexed header of HTTP request.

-> IO ByteString
source from client, for raw response

-> Response
HTTP response including status code and response header.

-> IO Bool
Returing True if the connection is persistent.

Sending a HTTP response to Connection according to Response.

Applications/middlewares MUST specify a proper ResponseHeaders. so that inconsistency does not happen. No header is deleted by this function.

Especially, Applications/middlewares MUST take care of Content-Length, Content-Range, and Transfer-Encoding because they are inserted, when necessary, regardless they already exist. This function does not insert Content-Encoding. It's middleware's responsibility.

The Date and Server header is added if not exist in HTTP response header.

There are three basic APIs to create Response:

responseFile :: Status -> ResponseHeaders -> FilePath -> Maybe FilePart -> Response

HTTP response body is sent by sendfile() for GET method. HTTP response body is not sent by HEAD method. Applications are categorized into simple and sophisticated. Simple applications should specify Nothing to Maybe FilePart. The size of the specified file is obtained by disk access. Then Range is handled. Sophisticated applications should specify Just to Maybe FilePart. They should treat Range (and If-Range) by themselves. In both cases, Content-Length and Content-Range (if necessary) are automatically added into the HTTP response header. If Content-Length and Content-Range exist in the HTTP response header, they would cause inconsistency. Status is also changed to 206 (Partial Content) if necessary.

responseBuilder :: Status -> ResponseHeaders -> Builder -> Response

HTTP response body is created from Builder. Transfer-Encoding: chunked is used in HTTP/1.1.

responseStream :: Status -> ResponseHeaders -> StreamingBody -> Response

HTTP response body is created from Builder. Transfer-Encoding: chunked is used in HTTP/1.1.

responseRaw :: (IO ByteString -> (ByteString -> IO ()) -> IO ()) -> Response -> Response

No header is added and no Transfer-Encoding: is applied.